fix: horizontal bottom sheets not being fully expanded
