--- core/ui/base/src/main/kotlin/com/f0x1d/logfox/core/ui/base/dialog/BaseBottomSheetFragment.kt
@@ -1,71 +1,0 @@
-package com.f0x1d.logfox.core.ui.base.dialog
-
-import android.annotation.SuppressLint
-import android.app.Dialog
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.lifecycle.Lifecycle
-import androidx.lifecycle.lifecycleScope
-import androidx.lifecycle.repeatOnLifecycle
-import androidx.viewbinding.ViewBinding
-import com.f0x1d.logfox.core.ui.base.ext.enableEdgeToEdge
-import com.google.android.material.bottomsheet.BottomSheetBehavior
-import com.google.android.material.bottomsheet.BottomSheetDialog
-import com.google.android.material.bottomsheet.BottomSheetDialogFragment
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.FlowCollector
-import kotlinx.coroutines.launch
-
-abstract class BaseBottomSheetFragment<T : ViewBinding> : BottomSheetDialogFragment() {
-
-    private var mutableBinding: T? = null
-    protected val binding: T get() = mutableBinding!!
-
-    abstract fun inflateBinding(inflater: LayoutInflater, container: ViewGroup?): T?
-    protected open fun T.onViewCreated(view: View, savedInstanceState: Bundle?) = Unit
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?,
-    ): View? {
-        inflateBinding(inflater, container)?.also {
-            mutableBinding = it
-            return it.root
-        }
-        return super.onCreateView(inflater, container, savedInstanceState)
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        binding.onViewCreated(view, savedInstanceState)
-    }
-
-    @SuppressLint("RestrictedApi")
-    override fun onCreateDialog(savedInstanceState: Bundle?): Dialog {
-        val dialog = super.onCreateDialog(savedInstanceState) as BottomSheetDialog
-        dialog.window?.enableEdgeToEdge(isContrastEnforced = false)
-        dialog.behavior.skipCollapsed = true
-        dialog.behavior.state = BottomSheetBehavior.STATE_EXPANDED
-        dialog.behavior.disableShapeAnimations() // i love google https://github.com/material-components/material-components-android/pull/437
-        return dialog
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        mutableBinding = null
-    }
-
-    protected fun <T> Flow<T>.collectWithLifecycle(
-        state: Lifecycle.State = Lifecycle.State.STARTED,
-        collector: FlowCollector<T>,
-    ) {
-        viewLifecycleOwner.lifecycleScope.launch {
-            viewLifecycleOwner.repeatOnLifecycle(state) {
-                collect(collector)
-            }
-        }
-    }
-}
